feat(Constant): 로또 1개당 가격 명시

+++ src/main/java/lotto/constant/LottoPrice.java
@@ -2,15 +2,15 @@ package lotto.constant;
 
 public enum LottoPrice {
 
-  LOTTO_PRICE(1000);
+  LOTTO_PRICE(1000L);
 
-  private final Integer price;
+  private final Long price;
 
-  LottoPrice(Integer price) {
+  LottoPrice(Long price) {
     this.price = price;
   }
 
-  public Integer getPrice() {
+  public Long getPrice() {
     return price;
   }
 }
